# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Livia API — multi-tenant appointment operating system
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/{businessId}/services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services for a business
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/BusinessId"
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/BusinessId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceBody"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /businesses/{businessId}/services/{serviceId}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a service
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/BusinessId"
        - $ref: "#/components/parameters/ServiceId"
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/BusinessId"
        - $ref: "#/components/parameters/ServiceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceBody"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Deactivate a service
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/BusinessId"
        - $ref: "#/components/parameters/ServiceId"
      responses:
        "204":
          description: Deactivated
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Customers ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalOpsSecret:
      type: apiKey
      in: header
      name: X-Internal-Ops-Secret
      description: Livia Inc operator secret (INTERNAL_OPS_SECRET or INTERNAL_CRON_SECRET fallback)
  parameters:
    BusinessId:
      name: businessId
      in: path
      required: true
      schema:
        type: string
...
    ServiceId:
      name: serviceId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Service:
      type: object
      properties:
        id:
          type: string
        businessId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        bufferBeforeMinutes:
          type: integer
        bufferAfterMinutes:
          type: integer
        priceMinor:
          type: integer
        currency:
          type: string
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, businessId, name, durationMinutes, bufferBeforeMinutes, bufferAfterMinutes, priceMinor, currency, isActive, sortOrder, createdAt, updatedAt]
    CreateServiceBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        durationMinutes:
          type: integer
        bufferBeforeMinutes:
          type: integer
          default: 0
        bufferAfterMinutes:
          type: integer
          default: 0
        priceMinor:
          type: integer
          default: 0
        currency:
          type: string
          default: GBP
        imageUrl:
          type: string
        sortOrder:
          type: integer
          default: 0
      required: [name, durationMinutes]
    UpdateServiceBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        durationMinutes:
          type: integer
        bufferBeforeMinutes:
          type: integer
        bufferAfterMinutes:
          type: integer
        priceMinor:
          type: integer
        currency:
          type: string
        imageUrl:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer

    # ── Customer ──
...